requirements.txt
# Web application
Flask>=3.0.0
gunicorn>=21.2.0
Werkzeug>=3.0.1
Jinja2>=3.1.2
requests>=2.31.0
python-dotenv>=1.0.0

# File handling
Pillow>=10.4.0

# Database (>=2.5 uses httpx>=0.25 which supports Python 3.14)

# ML / Analytics  (prophet is optional — code falls back to linear regression)
pandas>=2.2.3
numpy>=2.1.0
scikit-learn>=1.5.2

# Testing
pytest>=7.4.3
pytest-mock>=3.12.0
